+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Released: Upcoming
     - crudRenderFile
 - Added documentation about how to optimize serving the static content
 - Attention: Moved the events code up to the abstract data class which changed the signatures a bit
+- "before" "delete" Events of cascade deleted children are now properly taken into account
 
 ## 0.11.0
 Released: 2017-04-02